jasanjejop/requirements.txt
python-telegram-bot==21.9
playwright>=1.41.0
chromadb>=1.0.0
numpy>=2.0.0
openai>=1.50.0
python-dotenv==1.0.0
apscheduler==3.10.4
tavily-python
